services/reco-worker/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-httpx==0.47b0
